+++ CHANGELOG.md
@@ -14,5 +14,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   - Plurarization (`Intl.PluralRules`)
   - Numbers (`Intl.NumberFormat`)
   - Dates (`Intl.DateTimeFormat`)
+  - Lists (`Intl.ListFormat`)
 - `initI18n()` for initializing the internationalization
 - `dt()` for defining message translation